backend/requirements.txt
fastapi
uvicorn
python-multipart
boto3
opencv-python
pytesseract
python-dotenv
numpy
pydantic
httpx
Pillow
PyMuPDF
sqlalchemy
passlib[bcrypt]
PyJWT
email-validator
